+++ protected/humhub/assets/CoreApiAsset.php
@@ -8,14 +8,19 @@
 
 namespace humhub\assets;
 
-use humhub\components\assets\WebStaticAssetBundle;
+use humhub\components\assets\AssetBundle;
 use yii\web\View;
 
 /**
  * HumHub Core Api Asset
  */
-class CoreApiAsset extends WebStaticAssetBundle
+class CoreApiAsset extends AssetBundle
 {
+    /**
+     * @inheritdoc
+     */
+    public $sourcePath = '@humhub/resources';
+
     /**
      * @inheritdoc
      */